requirements.txt
flask
flask-session
python-dotenv
langchain
langchain-community
faiss-cpu
transformers
sentence-transformers
PyPDF2
